Split load more and .map in seperate components
