+++ src/Components/AllComics/AllComicsDetail/AllComicsDetail.scss
@@ -1,0 +1,34 @@
+@import "../../../SassVariables.scss";
+
+.AllComics {
+  .all-comics-img-container {
+    padding-top: 40px;
+    display: flex;
+    flex-wrap: wrap;
+    .all-comics-img {
+      margin-bottom: 40px;
+      padding-right: 4%;
+      max-width: 300px;
+      p {
+        color: $generalFontColor;
+        font-size: 1.1rem;
+        font-family: $fontStyleTwo;
+        padding-left: 3px;
+      }
+      .more {
+        position: relative;
+        margin: 0;
+        padding: 0;
+        font-size: 1.5rem;
+        font-weight: bold;
+        top: 200px;
+        left: 65px;
+        color: transparent;
+      }
+      img {
+        cursor: pointer;
+      }
+      @include imgHoverEffect;
+    }
+  }
+}
